Split audio generation server helpers
